cleaned up the aws_s3_helper spec
